pyproject.toml
[tool.poetry]
name = "commanderbot"
version = "0.20.0a23"
description = "A collection of utilities and extensions for discord.py bots."
authors = ["Arcensoth <arcensoth@gmail.com>"]
license = "MIT"
readme = 'README.md'
repository = "https://github.com/MinecraftCommands/commanderbot-py"
homepage = "https://github.com/MinecraftCommands/commanderbot-py"
keywords = ['discord', 'bot']

[tool.poetry.dependencies]
python = "^3.11"

# Core dependencies.
"discord.py" = "^2.3.0"
python-dotenv = "^1.0.0"

# These are used for logging.
colorama = "^0.4.0"
colorlog = "^6.7.0"

# These are used for managing SQLite databases.
aiosqlite = "^0.17.0"
# NOTE [SQLAlchemy] See: https://github.com/python-poetry/poetry/issues/4402
# This was fixed in stable, but doesn't seem to be fixed in the beta branch.
SQLAlchemy = ">=1.4,<1.4.23"

# These are used for handling text-based data (json, yaml, etc).
jsonpath-ng = "^1.6.0"
PyYAML = "^6.0"

# The rest are for various extensions.
aiohttp = "^3.9.0"
emoji = "^2.0.0"
allay = "^1.3.0"
mccq = "^1.0.2"
psutil = "^5.9.0"
feedparser = "^6.0.11"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"

[tool.isort]
profile = "black"

[tool.pyright]
typeCheckingMode = "basic"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
